feat: refactor landing page and routing logic
