+++ src/components/header.js
@@ -5,13 +5,13 @@ import { useRouter, useSearchParams } from "next/navigation";
 import Image from "next/image";
 import Logo from "../public/images/logo.png";
 import { NavMenu } from "./navMenu";
+import { useLocalStorage } from "@uidotdev/usehooks";
 
-export default function GlobalHeader({
-  userName,
-  handleUsernameChange,
-  setUserState,
-  room,
-}) {
+export default function GlobalHeader({ room }) {
+  const searchParams = useSearchParams();
+  const roomId = searchParams.get("roomid");
+  const [userName, setUserName] = useLocalStorage("userName", "");
+
   const { push } = useRouter();
   // const getInitials = (userName) => {
   //   return userName
@@ -20,8 +20,7 @@ export default function GlobalHeader({
   //     .toUpperCase();
   // };
   const handleLogoClick = () => {
-    setUserState({ userName: "", forcedLandingPage: true });
-    push("/");
+    push(`/${roomId ? `?roomid=${roomId}` : ""}`);
   };
 
   return (
