+++ src/components/participants.js
@@ -1,52 +1,55 @@
 "use client";
 
-import { useState } from "react";
 import classNames from "classnames";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+import { v4 as uuid4 } from "uuid";
+import { useLocalStorage } from "@uidotdev/usehooks";
+import { useSearchParams } from "next/navigation";
 
-export default function ParticipantsCard({
-  handleSetUserName,
-  isAdmin,
-  placeholderUserName = "",
-}) {
-  const [userName, setUserName] = useState(placeholderUserName);
+export default function ParticipantsCard() {
+  const searchParams = useSearchParams();
+  const roomId = searchParams.get("roomid");
+  const [userName, setUserName] = useLocalStorage("userName", "");
+
   const [isError, setError] = useState(false);
+  const { push } = useRouter();
 
   const handleSubmit = (e) => {
+    const newUserName = document.getElementById("user-name").value;
     e.preventDefault();
-    if (userName.length < 2) {
+    if (newUserName.length < 2) {
       setError(true);
       return;
     }
-    handleSetUserName(userName);
+    setUserName(newUserName);
+    push(`/room/?roomid=${roomId ?? uuid4()}`);
   };
   return (
     <>
       <form onSubmit={handleSubmit} className="flex flex-col gap-4">
         <div>
-          {isAdmin ? (
-            <>
-              <h1 className="text-lg font-bold">Create a Room</h1>
-            </>
-          ) : (
+          {roomId ? (
             <>
               <h1 className="text-lg text-primary font-bold">Join a room</h1>
               <h1 className="">{`You have been invited to join a room`}</h1>
+            </>
+          ) : (
+            <>
+              <h1 className="text-lg font-bold">Create a Room</h1>
             </>
           )}
         </div>
         <div className="form-control">
           <input
-            onChange={(e) => {
-              setUserName(e.target.value);
-            }}
-            value={userName}
+            id="user-name"
             type="text"
             autoFocus
-            placeholder="Enter your name"
+            placeholder={userName || "Enter your name"}
             className={classNames(
               "input input-bordered border border-primary",
               {
-                "border-red-600": isError && userName.length < 3,
+                "border-red-600": isError,
               },
             )}
           />
@@ -56,7 +59,7 @@ export default function ParticipantsCard({
             className="btn btn-primary text-gray-950 normal-case"
             type="submit"
           >
-            {isAdmin ? "Create Room" : "Join Room"}
+            {roomId ? "Join Room" : "Create Room"}
           </button>
         </div>
       </form>
